Separate backends and add backend selection options
